Support multiple player lives
Allow having more than one life for the player with the moddable default
set to 3. In a slight change, when the player contacts an enemy, it
loses a life in addition to being reset to the original position.
Similarly, falling into a danger zone resets the position in addition to
losing a life.

The lives are indicated by a row of hearts in the HUD. For simplicity, a
TextureRect that tiles the heart on expansion is used. It would probably
be more flexible to use a GridContainer.

https://phabricator.endlessm.com/T35581