Player: Implement jump buffering
Previously, you could only jump if you were already on the ground when you
press the jump button. This seems logical but is actually very annoying: if you
press jump 1 frame too early, then you miss.

It is common in platformers to allow a few frames' grace, so that you can press
the jump key a few frames early and still start a jump when the player touches
the ground. To implement this, count down from 0.1s (6 ticks at the default 60
ticks per second) when the jump action is just pressed.

This technique is borrowed from MOVEMENT 2.