Add MultiplayerCamera node
Add a new camera that inherits from Camera2D. It can frame the player
characters by positioning and zooming accordingly.

It adds a margin of 128 px (one tile size) to the player positions.

Add a new node for this camera to the main scene. It is disabled by
default. So by enabling this one and disabling or removing the current
camera for the character, this one takes precedence. The new multiplayer
camera has the same limits and position smoothing that the current
camera.

This is based on the Godot recipe [1] but adapted to Godot 4 and to this
project, with variable names improved.

[1] https://kidscancode.org/godot_recipes/3.x/2d/multi_target_camera/

https://github.com/endlessm/moddable-platformer/issues/9