Map joypad button 0, not joystick up, to jump
As a joypad user I would never expect to jump with the Up button on the
D-pad. And if you're using the joystick it's easy to accidentally
trigger the vertical axis and jump by mistake.

Remap the controls as follows:

- Rename the player_{1,2}_up actions in the input map to
  player_{1,2}_jump which more accurately reflects the meaning of the
  action.
- Remove the unused player_{1,2}_down actions.
- Map the bottom action button (Sony ×, Xbox A, Nintendo B) of each
  joypad to the corresponding _jump action.
- Unmap the vertical axis of the joystick from each _jump action.

I decided to leave the D-pad Up button in place because it's harder to
trigger accidentally.