Add physical W/A/D keys to input map
On an AZERTY keyboard, common in French-speaking areas, the keys
corresponding to the W/A/D positions on a US keyboard are Z/Q/D. On a
Dvorak keyboard, commonly used by nerds who like to make their own lives
more difficult and regret the choice they made 20 years ago, the keys
are ,/A/E.

Add the physical keycodes to the input map. Leave the logical keys in
the map so that pressing the key labelled “W” (documented in the
on-screen instructions) still works.