hud: Make touch screen buttons semitransparent
Previously, an enemy could creep up on you by being underneath one of
the buttons. This was more of a problem in multiplayer games, where the
camera zooms out only as far as necessary to fit both players on the
screen. Of course, there aren't touch controls for the second player,
but the learner could always add them.

Set the buttons to 75% opacity.