Replace deprecated TileMap with TileMapLayer
Since commit ef989983, this project has required Godot 4.3. In Godot
4.3, TileMap is deprecated in favour of TileMapLayer, and a warning to
that effect is shown in the scene tree.

The warning, and the documentation, explains how to migrate to
TileMapLayer. It was a little more involved than clicking that button
because we have a separate .tscn file for the tilemap, instance it in
the main scene, and draw the actual map in the instance in the main.
Essentially I used the automated process, then moved the node &
properties to the right places by hand.

https://docs.godotengine.org/en/stable/classes/class_tilemap.html