Improve input prompt overlay
When the game first starts, an overlay is shown with instructions
describing the controls. This was previously purely text-based.

Add assets for the controls from the Kenney Input Prompts pack
<https://kenney.nl/assets/input-prompts> (CC-0): arrow keys, W A S D,
and joypad buttons, for which I chose the Steam Deck assets which are
reasonably generic.

Rework the HUD to lay out the controls in a grid. Pack the keyboard
images and controller images into a TabContainer (with the tabs hidden)
which can be used to switch between the two sets of controls. Monitor
the Input.joy_connection_changed signal to switch between the two sets
of images based on the corresponding controller being connecter or
disconnected. This signal is emitted at the start of the game for each
connected controller, so we can default to the keyboard tab in each
case.

It might be nice to detect whether the touch controls are visible
and hide the control overlay in that case. This can be done by checking
DisplayServer.is_touchscreen_available(), which is what the touch
buttons use too. But this isn't a regression: we previously showed
keyboard controls even when touch controls were visible.

It might also be nice to only show the player two controls if there is
a player two. But again this is not a regression.