Workflows: Download and cache Godot Engine for build
Rather than relying on downloading and executing in a third-party container, we can just grab Godot Engine and the export templates directly from the upstream release page.

We cache these downloads so future runs have the binaries and templates (the latter of which are quite large) available right away for the run.