Global: Suppress UNUSED_SIGNAL warning
gravity_changed is emitted from GameLogic. Suppress the warning about it
not being used in global.gd and add a documentation comment explaining
it.

(The name of this signal is a little misleading because, in practice,
GameLogic sets the gravitational force exactly once in _ready(). But you
could imagine someone adding a way to vary gravity over the course of
the game, and the game elements that listen to this signal would behave
correctly.)