global: Pause timer rather than start/stopping
In the original implementation in this PR, the timer in the HUD would
not be displayed until you start moving, and would reset to 0.0 when you
win.

Instead, start the timer as soon as it is added, but pause it so that it
does not count down. Unpause it when the game starts. Pause it when the
game is won, so that the remaining time is displayed when you win.