player: Invert jump velocity parameter & adjust range
When the player jumps, the y component of their velocity is set to
jump_velocity. In Godot, up is negative. So velocity.y must be set to a
negative number.

But I don't think this means that the jump_velocity parameter needs to
be negative. And certainly it doesn't make sense for its range to
include jumping *downwards* by setting it to a positive number: I have
tested this and as expected it doesn't actually work.

Invert jump_velocity: set velocity.y to its negation when jumping.
Adjust the default value accordingly. Change the range to 0–2000 rather
than -1000–1000.