Simplify handling both players' inputs
The player-character scene can be configured to respond to player 1's
inputs, player 2's inputs, or both. I don't really know why you would
want it to respond to both, but it's there.

Previously the both case was implemented by checking the player 1 and
player 2 versions of each action, and combining them.

I think it is cleaner to have a "both" version of each action, bound to
the union of the player 1 and player 2 input events, and then just check
that directly.

Move all this handling to a new Actions global. Set up the "both"
actions dynamically (so that we don't have to manually keep them in sync
in the project settings).