player: Add teleport/blink action
If the player enters this action while also holding left or right, the
player-character teleports by a constant distance horizontally.

On keyboard, this is bound to "Enter" for player 1 (who moves with the
arrow keys), and "Shift" for player 2 (who moves with WASD).

On gamepad, this is bound to Xbox button B.

The function call for this ability is in the correct place in
_physics_process() but is commented out. A learner must find this
commented-out line and uncomment it to enable this ability.

The sound effect is https://freesound.org/people/D.S.G./sounds/244654/
(CC0-1.0).