player: Add "glide" special ability
If the player holds down the "jump" action, their character's downwards
speed is capped to a constant. This allows the player-character to
travel across wider gaps by jumping then gliding.

The function call for this ability is in _physics_process(), after other
"normal" adjustments to the velocity but before move_and_slide().
However, it is commented out. A learner must find this commented-out
line and uncomment it to enable this ability.

The sound effect is from
https://freesound.org/people/tothrec2/sounds/596541/, licensed under
CC0-1.0.