Build player-action mapping dynamically
Previously there was a constant dictionary mapping player enum × short
action name to full action name. This is very simple but is tedious to
keep updated as more actions are added.

Instead, build the mapping dynamically from a constant array of short
action names, and improve the actions.gd documentation.

As it stands this is a bit longer than writing it out by hand, but
future commits which will add new actions become one-liners.