Add Threadbare-themed tiles
To date this game has used generic assets (from Kenney's Platformer Art
collection). We now want the game to be Threadbare-themed.

Add new tiles and a corresponding tileset, identical to tileset-a.tres
except for the texture.

Visually, the tiles are *almost* interchangable with the 4 sets already
in the game. but the two variations of slopes are identical to each
other (unlike in the Kenney tiles, where one has straight lines and the
other wavy), and a couple of the others don't quite work.

But it is easy enough to adjust the scene to be compatible with either,
in a future commit.

I took the tiles from
https://github.com/endlessm/threadbare-origins-inkdrinker/ but modified
them to be more compatible with the Kenney tiles. They are in turn
derived from the same tiles in Threadbare. In Threadbare those assets
are under CC BY-SA 4.0 but since Endless Access owns the copyright on
them we can use them in this more liberally licensed project without
issue.