tilemap: Unset use_parent_material
The main scene's TileMapLayer is actually an instance of a tiny scene
which is an empty TileMapLayer node with its tile_set. I believe this is
so that the "reset" button beside the tile_set property on the instance
in the main scene gets you back to the default, rather than to a `null`
tileset (breaking the scene).

In commit 5222a3897761a75480ef90224b2c934909f8036c ("Replace deprecated
TileMap with TileMapLayer"), for unknown reasons I set
use_parent_material = true on the TileMapLayer scene. This was not
intentional, and it has no effect that I can see. Unset it.