platform: Make collision shapes local to scene
Previously, the collision shapes were shared between each instance of
the scene.

This means that if you placed 2 platforms in a scene, and set one of
them to be 2 tiles wide and the other to 3 tiles wide, the collision
shapes would both be as wide as the last platform in the scene tree,
rather than matching their visible size.

Making the resources local_to_scene causes them to be duplicated when
the scene is instanced.