Tweak platform under goal flag
The new Threadbare-themed tileset is almost, but not quite, compatible
with the Kenney tilesets. In cells (5, 2) and (6, 2), the Kenney
tilesets have "left and right edges of platform superimposed on
background", but the Threadbare tileset has "right and left (sic) edges
of half-tile platform with rock beneath" (i.e. more like (5, 1) and (6,
1), but with rock. The rocks don't join neatly to the grass-on-dirt
tiles used in the rest of the tileset.

But (5, 2) and (5, 3) are only used for the platform directly under the
flag, and even with the Kenney tiles they don't quite look right - they
don't join properly to the slope to the left.

Replace these with regular "top of platform" tiles, so that the tilesets
can be swapped without issue.