gitignore: Ignore Aseprite files
As noted in the comment, this project doesn't use Git LFS and I think
it's probably best to keep it that way. But while working on assets it's
really convenient to have the .aseprite file in the asset directory next
to the exported file(s).

Add such files to gitignore to prevent accidentally committing them.