platform: Add separate "fall" checkbox
Previously the fall_time property had a special case that, if the value is less
than zero, the platform would not fall. -1 was the default value. It was
exported with @export rather than @export_range, perhaps because the range would
have to include a range of negative numbers, all of which would behave the same.

Add a new `fall` boolean property, replacing the negative-time meaning of
`fall_time`. Group these two properties together in a "Falling Platform" group.
Make the `fall` property control whether the other property in the group is
visible at all. Change the `fall_time` property to use `@export_range`, and its
default value to 2.0.

Update the main scene to preserve the previous behaviour.