Replace crystal coins with Threadbare button coins
Starting with the [spinning button animation that Manuel drew for
Threadbare][0], I converted the yellow version to greyscale; scaled it
up to 128×128 (matching the existing coin assets in this project), and
completed the 7-frame animation. (In Threadbare, the frame order is 0 1
2 3 2* 1* 0* where the *s mean the frame is flipped horizontally; but
it's easier to use if it's a vanilla sprite sheet.)

Then I exported two assets:

- button_frames.png, containing the whole animation
- button.png, which has just the first frame

Then the coins are set to use button.png, keeping the yellow tint (i.e.
modulate).

button_frames.png is not used but it's in the source tree as an easter
egg to tempt learners to try to figure out how to change the coin to
be animated.

It's a little bit weird that we have collectible buttons and also
buttons stitched to the tiles, but for me these work better with the
surrounding game than stacks of books (as used in
threadbare-origins-inkdrinker) - we can keep the nice property of being
able to recolour the coins with tint/modulate.

I did however tweak the placement of some of the decorative coins in the
tilemap to compensate.

[0]: https://github.com/endlessm/threadbare/blob/f687d911c26ee31fe9f31b962d45a031f3860bb0/scenes/game_elements/props/button_item/components/button.aseprite