Add InkDrinkers from Threadbare as Player SpriteFrames
The InkDrinkers (minus shadow) are almost exactly the same height as the
Kenney player assets.

And by another happy coincidence, there is a 128×128px region inside the
full 192×192 frame of the InkDrinker which lines up perfectly with the
Kenney players AND cuts out the shadow. So I didn't even have to edit
the sprite sheets, just crop them appropriately in the SpriteFrames.

We don't quite have the right animations, so use the Threadbare idle
animations for the Moddable Platformer idle and jump_down, and
Threadbare move for Moddable Platformer walk and jump_up.