Replace enemy with Storyvore
For this one I couldn't use the sprite sheets from Threadbare directly,
because I had to remove the shadow under the feet. And so we only have
one colour. That's OK, we only had one colour before, too.

To preserve the ability to use the old enemy, duplicate the scene,
adjust the sprite in the duplicate, and change the main scene to use the
duplicate. The same script can be used by both scenes. This also allows
adjusting the position of the AnimatedSprite2D down a few pixels to make
the feet line up better with the ground.