Find coin nodes by their group
Previously, game_logic.gd walked recursively over all nodes in the tree,
counting those which are an instance of the Coin script.

A more idiomatic way to do this is to place coins into a group, and then
count how many nodes are in that group. This will also make it easier to
have some coins that don't count towards the goal, by having those coins
not be in this group.