Add SpecialCoin scene which is not in the coins group
This scene is identical to coin.tscn except that it is not in the
"coins" group. This will be used to add extra coins to the scene that
can't be collected with the default game setup: you have to give the
player additional abilities.

Annoyingly if scene A is in a group, then you can't remove instances of
A from that group; you also can't make an inherited scene B that
inherits from A and remove B from the group.

We could do it the other way round: have a basic coin which is not in
the group, and then make coin.tscn inherit from it just to set the
group. But that seems very strange.

Or we could not put coin.tscn into the group, and force the level
designer to add every non-special coin to the group by hand. This seems
error-prone.

Duplicating the scene seems the least-bad option; it's a very simple
scene, and the (simple) script is shared.