Update gamepad bindings for shrink & phase
Previously, the bindings were:

- Phase: Xbox X
- Shrink: DPad down, or press the left stick

I found these to be hard to use. It's hard to jump (A) while also
pressing X with the same thumb; and pushing in the stick is an
unfamiliar action. I tend to use the DPad but I am an old person; when I
see children pick up a controller, without fail they use the joystick.

Bind phase to R1 and shrink to L1 (or DPad down, still).