Replace individual input prompt assets with atlas
Previously we had just a few individual assets for the keys/buttons used
in the game.

We have added special actions and will want to (optionally) show them in
the HUD. Rather than having to import individual textures for each new
button, add a sprite sheet for keyboard and Steam Deck (a
generic-looking controller for our purposes), and create AtlasTexture
resources for each tile. Update the HUD scene and delete all the
previous assets.

This does mean a switch from SVG to PNG. Probably OK.

I used a script derived from
https://bitbucket.org/iseesharp83workspace/kenney-spritesheet-importer-godot-plugin
to generate the AtlasTexture resources but I had to fix some bugs in
that script, so ended up not adding the plugin.