HUD: Optionally show buttons for specials
If show_specials is true, show the keyboard/gamepad hints for shrink,
teleport, and phase. (Glide is just holding down jump, so does not need
a separate hint.) This is off by default because these actions need to
be uncommented to be used.

This expanded HUD is not particularly self-explanatory but perhaps
that's an incentive for learners to improve it!