player: Remove space between `#` and `_phase()`
Previously the four special abilities were commented out in the standard
way, with a space after the `#`:

	# _phase()

If you have no prior coding experience you may expect that it is enough
to delete the `#`. However that will cause a syntax error. You have to
delete the space as well.

So: remove the spaces.