Add "ding" sound when collecting coins
The sound effect is the third section of Coin flip.m4a by eZZin
<https://freesound.org/s/545857/>. Thanks to E.T. for helping me to
choose the sound.

License: Creative Commons 0

Previously, the coin scene would be freed at the end of the frame in
which it was collected. To allow the stream to play to completion,
instead hide the coin and disable collisions; then play the sound and
await its completion; and finally free the scene.