doc: Update screenshot
This is a recreation of the previous screenshot for the current version of the
game. The game looks quite different after being reskinned with Threadbare
assets.

The old screenshot was a PNG, but because there is more detail in the new
assets, the resulting PNG was 1.5 MB. Converting it to lossy WebP made it around
half the size of the previous screenshot and, to my eye, it looks perfectly
good.

(The Godot Asset Library supports WebP screenshots.)